+++ CHANGES.rst
@@ -15,6 +15,8 @@ Changelog of lizard-ui
 - Commented line 413 in jquery-ui-1.8.5.custom.css to remove conflicing
   (with Twitter Bootstrap) class .ui-button-text-only .ui-button-text
   [Gijs Nijholt]
+  
+- Fixed some interaction issues with modal and non-modal login form.
 
 
 4.0a2 (2012-05-18)